recs-service/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
implicit>=0.6.1
pydantic>=1.10.0
python-dotenv>=1.0.0
joblib>=1.2.0
supabase>=1.0.0
typing_extensions>=4.5.0r